descriptions can now be added but the height is wrong due to the fact that we need to size our widget to the width of the target before actually laying it out

+++ lib/classes_and_enums/dropdown_max_height.dart
@@ -4,7 +4,7 @@ class DropdownMaxHeight {
   ///of three rows. The user will have to scroll if the list grows larger.
   ///
   /// This will be ignored if byPixels is provided
-  final double byRows;
+  final int byRows;
 
   ///Define the max height of the dropdown using explicit pixels, for example, byPixels = 300,
   ///and the dropdown won't grow taller than 300 pixels
